Remove model-verifier code and move GradleProject to plugin-base.
