+++ .idea/.name
@@ -1,1 +1,1 @@
-spine-base
+spine
