+++ gradle-plugins/model-compiler/src/main/java/io/spine/gradle/compiler/ModelCompilerPlugin.java
@@ -24,7 +24,6 @@ import io.spine.gradle.compiler.cleaning.CleaningPlugin;
 import io.spine.gradle.compiler.lookup.enrichment.EnrichmentLookupPlugin;
 import io.spine.gradle.compiler.lookup.proto.ProtoToJavaMapperPlugin;
 import io.spine.gradle.compiler.lookup.valrule.ValidationRulesLookupPlugin;
-import io.spine.gradle.compiler.model.ModelVerifierPlugin;
 import io.spine.gradle.compiler.protoc.ProtocPluginImporter;
 import io.spine.gradle.compiler.rejection.RejectionGenPlugin;
 import io.spine.gradle.compiler.validate.ValidatingBuilderGenPlugin;
@@ -70,9 +69,6 @@ public class ModelCompilerPlugin implements Plugin<Project> {
 
         log().debug("Applying Spine protoc-plugin importer plugin.");
         new ProtocPluginImporter().apply(project);
-
-        log().debug("Applying Spine model verifier plugin.");
-        new ModelVerifierPlugin().apply(project);
     }
 
     private static Logger log() {
