+++ gradle-plugins/model-compiler/src/test/java/io/spine/gradle/compiler/rejection/RejectionGenPluginShould.java
@@ -23,8 +23,7 @@ package io.spine.gradle.compiler.rejection;
 import com.sun.javadoc.ClassDoc;
 import com.sun.javadoc.ConstructorDoc;
 import com.sun.javadoc.RootDoc;
-import io.spine.gradle.compiler.GradleProject;
-import org.junit.Ignore;
+import io.spine.gradle.GradleProject;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.rules.TemporaryFolder;
