+++ gradle-plugins/plugin-base/build.gradle
@@ -34,4 +34,5 @@ dependencies {
 
     testCompile group: 'junit', name: 'junit', version: jUnitVersion
     testCompile group: 'org.mockito', name: 'mockito-core', version: '1.9.0'
+    testCompile gradleTestKit()
 }
